+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTableHeaderRenderer.java
@@ -14,137 +14,51 @@
  * You should have received a copy of the GNU General Public License
  * along with Codion.  If not, see <https://www.gnu.org/licenses/>.
  *
- * Copyright (c) 2023 - 2025, Björn Darri Sigurðsson.
+ * Copyright (c) 2025, Björn Darri Sigurðsson.
  */
 package is.codion.swing.common.ui.component.table;
 
-import is.codion.common.model.condition.ConditionModel;
-import is.codion.swing.common.model.component.table.FilterTableSort.ColumnSortOrder;
+import is.codion.common.utilities.property.PropertyValue;
+import is.codion.swing.common.model.component.table.FilterTableModel;
 
-import org.jspecify.annotations.Nullable;
-
-import javax.swing.AbstractButton;
-import javax.swing.Icon;
-import javax.swing.JLabel;
-import javax.swing.JTable;
-import javax.swing.SortOrder;
-import javax.swing.border.Border;
 import javax.swing.table.TableCellRenderer;
-import java.awt.Color;
-import java.awt.Component;
-import java.awt.Font;
-import java.awt.Graphics;
 
-import static javax.swing.BorderFactory.createCompoundBorder;
+import static is.codion.common.utilities.Configuration.booleanValue;
 
-final class FilterTableHeaderRenderer<R, C> implements TableCellRenderer {
+/**
+ * A renderer for a {@link FilterTable} header.
+ */
+public interface FilterTableHeaderRenderer extends TableCellRenderer {
 
-	private static final int SORT_ICON_SIZE = 5;
+	/**
+	 * Specifies whether the focused column should be indicated with a darkened header.
+	 * <ul>
+	 * <li>Value type: Boolean
+	 * <li>Default value: true
+	 * </ul>
+	 */
+	PropertyValue<Boolean> FOCUSED_COLUMN_INDICATOR =
+					booleanValue(FilterTableHeaderRenderer.class.getName() + ".focusedColumnIndicator", true);
 
-	private final FilterTable<R, C> filterTable;
-	private final TableCellRenderer wrappedRenderer;
-	private final TableCellRenderer columnCellRenderer;
+	/**
+	 * A factory for {@link FilterTableHeaderRenderer} instances.
+	 * @param <C> the column identifier type
+	 */
+	interface Factory<C> {
 
-	FilterTableHeaderRenderer(FilterTable<R, C> filterTable, FilterTableColumn<C> column) {
-		this.filterTable = filterTable;
-		this.wrappedRenderer = column.getHeaderRenderer();
-		this.columnCellRenderer = column.getCellRenderer();
+		/**
+		 * @param column the column
+		 * @param tableModel the table model
+		 * @return a {@link FilterTableHeaderRenderer} instance for the given column
+		 */
+		FilterTableHeaderRenderer create(FilterTableColumn<C> column, FilterTableModel<?, C> tableModel);
 	}
 
-	@Override
-	public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected,
-																								 boolean hasFocus, int row, int column) {
-		Component component = wrappedRenderer == null ?
-						filterTable.getTableHeader().getDefaultRenderer()
-										.getTableCellRendererComponent(filterTable, value, isSelected, hasFocus, row, column) :
-						wrappedRenderer.getTableCellRendererComponent(filterTable, value, isSelected, hasFocus, row, column);
-		Font defaultFont = component.getFont();
-		if (component instanceof JLabel) {
-			JLabel label = (JLabel) component;
-			FilterTableColumn<C> tableColumn = filterTable.columnModel().getColumn(column);
-			ConditionModel<?> filterModel = filterTable.model().filters().optional(tableColumn.identifier()).orElse(null);
-			label.setFont((filterModel != null && filterModel.enabled().is()) ? defaultFont.deriveFont(Font.ITALIC) : defaultFont);
-			label.setIcon(sortArrowIcon(tableColumn.identifier(), label.getFont().getSize() + SORT_ICON_SIZE));
-			label.setIconTextGap(0);
-			if (columnCellRenderer instanceof JLabel) {
-				label.setHorizontalAlignment(((JLabel) columnCellRenderer).getHorizontalAlignment());
-			}
-			else if (columnCellRenderer instanceof AbstractButton) {
-				label.setHorizontalAlignment(((AbstractButton) columnCellRenderer).getHorizontalAlignment());
-			}
-			if (columnCellRenderer instanceof DefaultFilterTableCellRenderer) {
-				Border tableCellBorder = ((DefaultFilterTableCellRenderer<?, ?, ?>) columnCellRenderer).settings().defaultCellBorder();
-				label.setBorder(label.getBorder() == null ? tableCellBorder : createCompoundBorder(label.getBorder(), tableCellBorder));
-			}
-		}
-
-		return component;
-	}
-
-	private @Nullable Icon sortArrowIcon(C identifier, int iconSizePixels) {
-		ColumnSortOrder<C> columnSortOrder = filterTable.model().sort().columns().get(identifier);
-
-		return columnSortOrder.sortOrder() == SortOrder.UNSORTED ? null : new Arrow(columnSortOrder.sortOrder() == SortOrder.DESCENDING, iconSizePixels, columnSortOrder.priority());
-	}
-
-	private static final class Arrow implements Icon {
-
-		private static final double PRIORITY_SIZE_RATIO = 0.8;
-		private static final double PRIORITY_SIZE_CONST = 2.0;
-		private static final int ALIGNMENT_CONSTANT = 6;
-
-		private final boolean descending;
-		private final int size;
-		private final int priority;
-
-		private Arrow(boolean descending, int size, int priority) {
-			this.descending = descending;
-			this.size = size;
-			this.priority = priority;
-		}
-
-		@Override
-		public void paintIcon(Component c, Graphics g, int x, int y) {
-			Color color = c == null ? Color.GRAY : c.getBackground();
-			// In a compound sort, make each successive triangle 20% smaller than the previous one.
-			int dx = (int) (size / PRIORITY_SIZE_CONST * Math.pow(PRIORITY_SIZE_RATIO, priority));
-			int dy = descending ? dx : -dx;
-			// Align icon (roughly) with font baseline.
-			int theY = y + SORT_ICON_SIZE * size / ALIGNMENT_CONSTANT + (descending ? -dy : 0);
-			int shift = descending ? 1 : -1;
-			g.translate(x, theY);
-
-			// Right diagonal.
-			g.setColor(color.darker());
-			g.drawLine(dx / 2, dy, 0, 0);
-			g.drawLine(dx / 2, dy + shift, 0, shift);
-
-			// Left diagonal.
-			g.setColor(color.brighter());
-			g.drawLine(dx / 2, dy, dx, 0);
-			g.drawLine(dx / 2, dy + shift, dx, shift);
-
-			// Horizontal line.
-			if (descending) {
-				g.setColor(color.darker().darker());
-			}
-			else {
-				g.setColor(color.brighter().brighter());
-			}
-			g.drawLine(dx, 0, 0, 0);
-
-			g.setColor(color);
-			g.translate(-x, -theY);
-		}
-
-		@Override
-		public int getIconWidth() {
-			return size;
-		}
-
-		@Override
-		public int getIconHeight() {
-			return size;
-		}
+	/**
+	 * @param <C> the column identifier type
+	 * @return a {@link FilterTableHeaderRenderer.Factory} instance
+	 */
+	static <C> FilterTableHeaderRenderer.Factory<C> factory() {
+		return (Factory<C>) DefaultFilterTableHeaderRenderer.FACTORY;
 	}
 }